Add support for icons
GNTP allows attaching icons to applications and notifications. When
making a growler, registering a notification, or sending a notification
an icon can be added as an optional argument. When specified as a URL,
the URL for the icon is added. When specified as a File, the File data
is directly embedded.