Add support for callbacks
GNTP supports callbacks on notifications. A callback is triggered when a
user clicks on or closes a notification, or it times out. If a callback
target is specified as a URL, that URL is opened in the user's browser
when the notification is clicked. Otherwise, the application can send a
callback context and type header, which are echoed back when the
callback is triggered.