Allow replacing notifications
GNTP allows attaching an id to a notification that matches a previous
notification. This serves as a hint to the notification server that the
new notification should replace/update the existing notification. The
spec specifies that this coalescing id will replace a notification with
a matching id. However, according to this
[bug report](http://code.google.com/p/growl/issues/detail?id=381), both
the Growl on Mac and the one on Windows only replace the notification if
both notifications having matching coalescing ids. So by default, we
send matching ids, and when supplied with a replaces id, we use that one
as the coalescing id.