Create version files
In preparation for adding the optional Origin headers to GNTP requests,
we need to create resources necessary to define a project's version, and
to refer to the version from code.

My lein-ver plugin was created for just this purpose. Therefore, we have
run `lein ver init` in order to generate the necessary resources.